Updates set up logic for Drupal 8.

+++ src/Robo/Drupal/BaseDrupalD8.php
@@ -12,12 +12,16 @@ class BaseDrupalD8 extends BaseDrupal {
    */
   public function setup() {
     // If .htaccess file or local settings file does not exit lets create them.
-    if (!file_exists($this->getSiteRoot() . '.htaccess') || !file_exists($this->getSiteRoot() . 'sites/default/settings.local.php')) {
+    if ((file_exists($this->getSiteRoot() . '.htaccess.default') && !file_exists($this->getSiteRoot() . '.htaccess')) || !file_exists($this->getSiteRoot() . 'sites/default/settings.local.php')) {
       $this->say("Missing .htaccess or settings.local.php");
       $name = $this->confirm("Missing .htaccess or settings.local.php Copy the default?");
       if ($name) {
-        $this->_exec('cp ' . $this->getSiteRoot() . '.htaccess.default ' . $this->getSiteRoot() . '.htaccess');
-        $this->_exec('cp ' . $this->getSiteRoot() . 'sites/example.settings.local.php ' . $this->$this->getSiteRoot() . 'sites/default/settings.local.php');
+        // Only copy the .htaccess.default if it default exists.
+        if (file_exists($this->getSiteRoot() . '.htaccess.default')) {
+          $this->_exec('cp ' . $this->getSiteRoot() . '.htaccess.default ' . $this->getSiteRoot() . '.htaccess');
+        }
+
+        $this->_exec('cp ' . $this->getSiteRoot() . 'sites/default/example.settings.local.php ' . $this->getSiteRoot() . 'sites/default/settings.local.php');
         $this->npmInstall();
         $this->siteInit = TRUE;
         $this->start();
